+++ src/main/java/com/srltas/runtogether/application/exception/BusinessException.java
@@ -1,5 +1,6 @@
-package com.srltas.runtogether.common.exception;
+package com.srltas.runtogether.application.exception;
 
+import com.srltas.runtogether.common.exception.RunTogetherException;
 import com.srltas.runtogether.common.exception.code.ErrorCode;
 
 public abstract class BusinessException extends RunTogetherException {
