+++ src/main/java/com/srltas/runtogether/domain/model/neighborhood/exception/NeighborhoodException.java
@@ -1,6 +1,6 @@
 package com.srltas.runtogether.domain.model.neighborhood.exception;
 
-import com.srltas.runtogether.common.exception.BusinessException;
+import com.srltas.runtogether.application.exception.BusinessException;
 import com.srltas.runtogether.common.exception.code.ErrorCode;
 
 public class NeighborhoodException extends BusinessException {
