+++ src/test/java/com/srltas/runtogether/adapter/in/web/filter/AuthenticationFilterTest.java
@@ -19,7 +19,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import com.srltas.runtogether.adapter.out.session.SessionStorage;
 import com.srltas.runtogether.adapter.out.session.UserSessionDTO;
-import com.srltas.runtogether.common.exception.UnauthorizedException;
+import com.srltas.runtogether.application.exception.UnauthorizedException;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
